+++ jst-core/src/main/java/io/github/ieu/jst/inventory/JstQueryInventoryResponse.java
@@ -49,7 +49,7 @@ public class JstQueryInventoryResponse {
             /**
              * 时间戳
              */
-            private Integer ts;
+            private Long ts;
 
             /**
              * 款式编码
