+++ jst-core/src/main/java/io/github/ieu/jst/stockin/JstQueryPurchaseInResponse.java
@@ -59,7 +59,7 @@ public class JstQueryPurchaseInResponse {
             /**
              * 数据库行版本号：https://docs.microsoft.com/zh-cn/sql/t-sql/data-types/rowversion-transact-sql?view=sql-server-ver16
              */
-            private Integer ts;
+            private Long ts;
 
             /**
              * 仓库名称
